% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{tb_pal}
\alias{tb_pal}
\title{TB color palette}
\usage{
tb_pal(palette = "all", reverse = FALSE)
}
\arguments{
\item{palette}{Chose a palette from tb.color list. Defaoult is all}

\item{reverse}{logical. If True palette is reversed}
}
\description{
TB color palette
}
