% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\docType{data}
\name{tb_palettes}
\alias{tb_palettes}
\title{Complete list of palettes}
\format{An object of class \code{list} of length 1.}
\usage{
tb_palettes
}
\description{
Complete list of palettes
}
\keyword{datasets}
