% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_tb.R
\name{theme_tb}
\alias{theme_tb}
\title{TB Theme}
\usage{
theme_tb(base_size = 11, base_family = "",
  plot_title_family = base_family, base_line_size = base_size/21,
  base_rect_size = 10.5/21, major.linetype = 2)
}
\arguments{
\item{base_size}{base font size}

\item{base_family}{base font family}

\item{base_line_size}{base size for line elements}

\item{base_rect_size}{base size for rect elements}
}
\description{
A complete theme based no gplot2::theme_minimal
}
